[FIX] Fix eeprom_update_block multiple definition link error
When compiling for AVR targets, the following error occurs:

    (.text+0x0): multiple definition of `eeprom_update_block'

Remove `eeprom_update_block` shim definition which is causing the error.

Notes
-----
Older versions of `avr-gcc` (like the one included with Arduino IDE 1.0.5)
do not include the `eeprom_update_block` function.  A shim `eeprom_write_block`
function was added to handle this scenario, but with current `avr-gcc` versions
this causes an error.