fix(conda): fix error due to existing directory
The `nanopb` package, which a required dependency of this package, writes
headers to the `share/platformio/include/nanopb` directory.  This package
also attempts to write headers to the `share/platformio/include/nanopb`
directory.  However, prior to this commit, this failed because the
`shutil.copytree` function does not support copying to a directory that
already exists.

In this commit, switch to use [`distutils.dir_util.copy_tree` instead][1]
since it supports copying to an existing directory.

[1]: https://stackoverflow.com/a/31039095/345236